refactor: Extract merge protocol tests to dedicated test file
Moved merge-related tests from mega_honk.test.cpp to a new dedicated
merge.test.cpp file in the goblin folder. This better organizes the test
suite and removes the goblin dependency from ultra_honk tests.

Changes:
- Created goblin/merge.test.cpp with merge-specific tests:
  * MergeProofSizeCheck
  * SingleCircuit
  * MultipleCircuitsMergeOnly
  * MultipleCircuitsHonkOnly
  * MultipleCircuitsMergeOnlyPrependThenAppend
  * MultipleCircuitsHonkAndMerge
  * OpQueueWithRandomValues

- Removed these tests from ultra_honk/mega_honk.test.cpp
- Removed construct_and_verify_merge_proof helper method from MegaHonkTests
- Removed merge-related includes from mega_honk.test.cpp
- Removed goblin dependency from ultra_honk CMakeLists.txt

All tests compile and link successfully.
