+++ barretenberg/cpp/src/barretenberg/ultra_honk/CMakeLists.txt
@@ -1,1 +1,1 @@
-barretenberg_module(ultra_honk goblin honk sumcheck commitment_schemes stdlib_primitives stdlib_keccak stdlib_sha256)
+barretenberg_module(ultra_honk honk sumcheck commitment_schemes stdlib_primitives stdlib_keccak stdlib_sha256)
