+++ barretenberg/cpp/src/barretenberg/ultra_honk/mega_honk.test.cpp
@@ -4,8 +4,6 @@
 
 #include "barretenberg/circuit_checker/circuit_checker.hpp"
 #include "barretenberg/common/log.hpp"
-#include "barretenberg/goblin/merge_prover.hpp"
-#include "barretenberg/goblin/merge_verifier.hpp"
 #include "barretenberg/goblin/mock_circuits.hpp"
 #include "barretenberg/honk/relation_checker.hpp"
 #include "barretenberg/stdlib_circuit_builders/mega_circuit_builder.hpp"
@@ -48,32 +46,6 @@ template <typename Flavor> class MegaHonkTests : public ::testing::Test {
 
         return verified;
     }
-
-    /**
-     * @brief Construct and verify a Goblin ECC op queue merge proof
-     *
-     */
-    bool construct_and_verify_merge_proof(auto& op_queue, MergeSettings settings = MergeSettings::PREPEND)
-    {
-        MergeProver merge_prover{ op_queue, settings };
-        auto merge_proof = merge_prover.construct_proof();
-
-        // Construct Merge commitments
-        MergeVerifier::InputCommitments merge_commitments;
-        auto t_current = op_queue->construct_current_ultra_ops_subtable_columns();
-        auto T_prev = op_queue->construct_previous_ultra_ops_table_columns();
-        for (size_t idx = 0; idx < Flavor::NUM_WIRES; idx++) {
-            merge_commitments.t_commitments[idx] = merge_prover.pcs_commitment_key.commit(t_current[idx]);
-            merge_commitments.T_prev_commitments[idx] = merge_prover.pcs_commitment_key.commit(T_prev[idx]);
-        }
-
-        auto transcript = std::make_shared<NativeTranscript>();
-        MergeVerifier merge_verifier{ settings, transcript };
-        auto [pairing_points, _, degree_check_passed, concatenation_check_passed] =
-            merge_verifier.verify_proof(merge_proof, merge_commitments);
-
-        return pairing_points.check() && degree_check_passed && concatenation_check_passed;
-    }
 };
 
 TYPED_TEST_SUITE(MegaHonkTests, FlavorTypes);
@@ -102,25 +74,6 @@ TYPED_TEST(MegaHonkTests, ProofLengthCheck)
     UltraProver_<Flavor> prover(prover_instance, verification_key);
     HonkProof mega_proof = prover.construct_proof();
     EXPECT_EQ(mega_proof.size(), Flavor::PROOF_LENGTH_WITHOUT_PUB_INPUTS() + DefaultIO::PUBLIC_INPUTS_SIZE);
-}
-
-/**
- * @brief Check that size of a merge proof matches the corresponding constant
- * @details This is useful for ensuring correct construction of mock merge proofs
- *
- */
-TYPED_TEST(MegaHonkTests, MergeProofSizeCheck)
-{
-    using Flavor = TypeParam;
-
-    auto builder = typename Flavor::CircuitBuilder{};
-    GoblinMockCircuits::construct_simple_circuit(builder);
-
-    // Construct a merge proof and ensure its size matches expectation; if not, the constant may need to be updated
-    MergeProver merge_prover{ builder.op_queue };
-    auto merge_proof = merge_prover.construct_proof();
-
-    EXPECT_EQ(merge_proof.size(), MERGE_PROOF_SIZE);
 }
 
 /**
@@ -202,144 +155,6 @@ TYPED_TEST(MegaHonkTests, DynamicVirtualSizeIncrease)
 }
 
 /**
- * @brief Test proof construction/verification for a circuit with ECC op gates, public inputs, and basic arithmetic
- * gates
- * @note We simulate op queue interactions with a previous circuit so the actual circuit under test utilizes an op queue
- * with non-empty 'previous' data. This avoid complications with zero-commitments etc.
- *
- */
-TYPED_TEST(MegaHonkTests, SingleCircuit)
-{
-    using Flavor = TypeParam;
-    auto builder = typename Flavor::CircuitBuilder{};
-
-    GoblinMockCircuits::construct_simple_circuit(builder);
-
-    // Construct and verify Honk proof
-    bool honk_verified = this->construct_and_verify_honk_proof(builder);
-    EXPECT_TRUE(honk_verified);
-
-    // Construct and verify Goblin ECC op queue Merge proof
-    auto merge_verified = this->construct_and_verify_merge_proof(builder.op_queue);
-    EXPECT_TRUE(merge_verified);
-}
-
-/**
- * @brief Test Merge proof construction/verification for multiple circuits with ECC op gates, public inputs, and
- * basic arithmetic gates
- *
- */
-TYPED_TEST(MegaHonkTests, MultipleCircuitsMergeOnly)
-{
-    using Flavor = TypeParam;
-    // Instantiate EccOpQueue. This will be shared across all circuits in the series
-    auto op_queue = std::make_shared<bb::ECCOpQueue>();
-    // Construct multiple test circuits that share an ECC op queue. Generate and verify a proof for each.
-    size_t NUM_CIRCUITS = 3;
-    for (size_t i = 0; i < NUM_CIRCUITS; ++i) {
-        auto builder = typename Flavor::CircuitBuilder{ op_queue };
-
-        GoblinMockCircuits::construct_simple_circuit(builder);
-
-        // Construct and verify Goblin ECC op queue Merge proof
-        auto merge_verified = this->construct_and_verify_merge_proof(op_queue);
-        EXPECT_TRUE(merge_verified);
-    }
-}
-
-TYPED_TEST(MegaHonkTests, MultipleCircuitsMergeOnlyPrependThenAppend)
-{
-    using Flavor = TypeParam;
-    // Instantiate EccOpQueue. This will be shared across all circuits in the series
-    auto op_queue = std::make_shared<bb::ECCOpQueue>();
-    // Construct multiple test circuits that share an ECC op queue. Generate and verify a proof for each.
-    size_t NUM_CIRCUITS = 3;
-    for (size_t i = 0; i < NUM_CIRCUITS; ++i) {
-        auto builder = typename Flavor::CircuitBuilder{ op_queue };
-
-        GoblinMockCircuits::construct_simple_circuit(builder);
-
-        // Construct and verify Goblin ECC op queue Merge proof
-        auto merge_verified = this->construct_and_verify_merge_proof(op_queue);
-        EXPECT_TRUE(merge_verified);
-    }
-
-    // Construct a final circuit and append its ecc ops to the op queue
-    auto builder = typename Flavor::CircuitBuilder{ op_queue };
-
-    GoblinMockCircuits::construct_simple_circuit(builder);
-
-    // Construct and verify Goblin ECC op queue Merge proof
-    auto merge_verified = this->construct_and_verify_merge_proof(op_queue, MergeSettings::APPEND);
-    EXPECT_TRUE(merge_verified);
-}
-
-/**
- * @brief Test Honk proof construction/verification for multiple circuits with ECC op gates, public inputs, and
- * basic arithmetic gates
- *
- */
-TYPED_TEST(MegaHonkTests, MultipleCircuitsHonkOnly)
-{
-    using Flavor = TypeParam;
-
-    // Instantiate EccOpQueue. This will be shared across all circuits in the series
-    auto op_queue = std::make_shared<bb::ECCOpQueue>();
-    // Construct multiple test circuits that share an ECC op queue. Generate and verify a proof for each.
-    size_t NUM_CIRCUITS = 3;
-    for (size_t i = 0; i < NUM_CIRCUITS; ++i) {
-        auto builder = typename Flavor::CircuitBuilder{ op_queue };
-        GoblinMockCircuits::construct_simple_circuit(builder);
-        // Construct and verify Honk proof
-        bool honk_verified = this->construct_and_verify_honk_proof(builder);
-        EXPECT_TRUE(honk_verified);
-        // Artificially merge the op queue sincer we're not running the merge protocol in this test
-        builder.op_queue->merge();
-    }
-}
-
-/**
- * @brief Test Honk and Merge proof construction/verification for multiple circuits with ECC op gates, public inputs,
- * and basic arithmetic gates
- *
- */
-TYPED_TEST(MegaHonkTests, MultipleCircuitsHonkAndMerge)
-{
-    using Flavor = TypeParam;
-
-    // Instantiate EccOpQueue. This will be shared across all circuits in the series
-    auto op_queue = std::make_shared<bb::ECCOpQueue>();
-    // Construct multiple test circuits that share an ECC op queue. Generate and verify a proof for each.
-    size_t NUM_CIRCUITS = 3;
-    for (size_t i = 0; i < NUM_CIRCUITS; ++i) {
-        auto builder = typename Flavor::CircuitBuilder{ op_queue };
-
-        GoblinMockCircuits::construct_simple_circuit(builder);
-
-        // Construct and verify Honk proof
-        bool honk_verified = this->construct_and_verify_honk_proof(builder);
-        EXPECT_TRUE(honk_verified);
-
-        // Construct and verify Goblin ECC op queue Merge proof
-        auto merge_verified = this->construct_and_verify_merge_proof(op_queue);
-        EXPECT_TRUE(merge_verified);
-    }
-
-    // Construct a final circuit whose ecc ops will be appended rather than prepended to the op queue
-    auto builder = typename Flavor::CircuitBuilder{ op_queue };
-
-    GoblinMockCircuits::construct_simple_circuit(builder);
-
-    // Construct and verify Honk proof
-    bool honk_verified = this->construct_and_verify_honk_proof(builder);
-    EXPECT_TRUE(honk_verified);
-
-    // Construct and verify Goblin ECC op queue Merge proof
-    auto merge_verified = this->construct_and_verify_merge_proof(op_queue, MergeSettings::APPEND);
-    EXPECT_TRUE(merge_verified);
-}
-
-/**
  * @brief A sanity check that a simple std::swap on a ProverPolynomials object works as expected
  * @details Constuct two valid proving keys. Tamper with the prover_polynomials of one key then swap the
  * prover_polynomials of the two keys. The key who received the tampered polys leads to a failed verification while the
@@ -394,40 +209,5 @@ TYPED_TEST(MegaHonkTests, PolySwap)
         auto proof = prover.construct_proof();
         bool result = verifier.template verify_proof<DefaultIO>(proof).result;
         EXPECT_FALSE(result);
-    }
-}
-
-/**
- * @brief To ensure the Merge proof sent to the rollup is ZK we have to randomise the commitments and evaluation of
- * column polynomials. We achieve this by adding some randomness to the op queue via random ops in builders. As we
- * produce a MegaHonk proof for such builders, this test ensure random ops do not alter the correctness of such proof
- * (which only asserts that the data in ecc_op_wires has been compied correctly from the other wires).
- *
- */
-TYPED_TEST(MegaHonkTests, OpQueueWithRandomValues)
-{
-    using Flavor = TypeParam;
-    using Builder = Flavor::CircuitBuilder;
-
-    // Test for randomness added at the beginning
-    {
-        Builder builder;
-        GoblinMockCircuits::randomise_op_queue(builder, 2);
-        GoblinMockCircuits::construct_simple_circuit(builder);
-
-        // Construct and verify Honk proof
-        bool honk_verified = this->construct_and_verify_honk_proof(builder);
-        EXPECT_TRUE(honk_verified);
-    }
-
-    // Test for randomness added at the end
-    {
-        Builder builder;
-        GoblinMockCircuits::construct_simple_circuit(builder);
-        GoblinMockCircuits::randomise_op_queue(builder, 2);
-
-        // Construct and verify Honk proof
-        bool honk_verified = this->construct_and_verify_honk_proof(builder);
-        EXPECT_TRUE(honk_verified);
     }
 }
